+++ src/rules/helper/has-alt-text.ts
@@ -1,0 +1,12 @@
+import { HtmlElement } from "../../dom";
+
+export function hasAltText(image: HtmlElement): boolean {
+	const alt = image.getAttribute("alt");
+
+	/* missing or boolean */
+	if (alt === null || alt.value === null) {
+		return false;
+	}
+
+	return alt.isDynamic || alt.value.toString() !== "";
+}