+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Features
 
+- new rule `wcag/h36`.
+- new rule `wcag/h30`.
+- new rule `long-title`.
+- new rule `empty-title`.
 - add `UserError` exception which is to be used for any error which is not
   caused by an internal error, e.g. configuration errors or a plugin. The error
   supresses the notice about internal error which should be reported as a bug.